compiler: reorg intermediate representation
