+++ crates/lean_compiler/src/b_compile_intermediate.rs
@@ -1,4 +1,4 @@
-use crate::{F, a_simplify_lang::*, intermediate_bytecode::*, lang::*, precompiles::*};
+use crate::{F, a_simplify_lang::*, ir::*, lang::*, precompiles::*};
 use lean_vm::*;
 use p3_field::Field;
 use std::{
