+++ crates/lean_compiler/src/lib.rs
@@ -10,7 +10,7 @@ use crate::{
 mod a_simplify_lang;
 mod b_compile_intermediate;
 mod c_compile_final;
-mod intermediate_bytecode;
+pub mod ir;
 mod lang;
 mod parser;
 mod precompiles;
